16 Dec
Renamed GameManager to GameManagerTF

Added 2 scripts to keep track of the button pressed as well as to resstore the answer if de-selected
(user press answer button -> displayed at first blank. This answer disappears from the answers option panel.
If user decides to change or remove this answer, the answer will re-appear in the answer option panel)
- attached both scripts to 2 prefabs respectively